Rename multiple components, remove some console logs
